requirements.txt
numpy==1.21.2
Pillow==8.4.0
torch==1.7.1
torchvision==0.8.2
tqdm==4.62.2
tensorboard==2.6.0
opencv==4.0.1
cudatoolkit==10.1.243
